Excludes: keep history of removed entries
So far the added and removed excludes have been in git but it makes it
hard to see what's the state. We want to remove all excludes eventually
but for future reference want to keep some nots.

Put them in to a separate file, same format as EXCLUDES so they can be
used eventually for testing as any other exclude files.

Workflow:
- check the date of last change in the exclude
- add entry to EXCLUDES.history, keep comment or explain why the exclude
  was added or removed
- remove the commits from git

Signed-off-by: David Sterba <dsterba@suse.com>