Workaround for systems without udev
When there's no udev daemon running, some of the dmsetup commands hang
and wait for events facilitated by udev. Like symlink creation or
deletion. When this started is not exactly known but it's a recent point
in time. The tests btrfs/056 or btrfs/085 can be used as an indicator
whether this works or not.

Due to the specific needs of the vmk environment with minimal init
setup, the following workaround is done in fstests:

- insert stub scripts in place of the expected dmsetup path
- udevadm is also stubbed, by udev_wait
- the script captures known problematic commands, create, resume, remove
- do not assume udev, skip waiting for the events
- substitute the actions that udev would, do namely mknodes, creating
  the device nodes in /dev/ or /dev/mapper (also removing them)
- the artificial delays in bin/dmsetup may not be necessary, the waiting
  is done by udev_wait script checkinf if the device node exists or not

Debugging tip: use _kernlog for tracing what's going on.

Signed-off-by: David Sterba <dsterba@suse.com>